Refactor Lib
