+++ src/API/Common/Interfaces/ICarryableTransfer.cs
@@ -1,7 +1,7 @@
 using Vintagestory.API.Common;
 using Vintagestory.API.Datastructures;
 
-namespace CarryOn.API.Common
+namespace CarryOn.API.Common.Interfaces
 {
     public interface ICarryableTransfer
     {
