+++ src/API/Common/Models/CarrySlot.cs
@@ -1,4 +1,4 @@
-namespace CarryOn.API.Common
+namespace CarryOn.API.Common.Models
 {
     /// <summary> Represents a location where something is or may be carried. </summary>
     public enum CarrySlot
