+++ src/API/Event/CarryEvents.cs
@@ -1,5 +1,5 @@
 using System;
-using CarryOn.API.Common;
+using CarryOn.API.Common.Models;
 using CarryOn.API.Event.Data;
 using Vintagestory.API.Common.Entities;
 using Vintagestory.API.MathTools;
